fix: pr-review findings
