+++ override-macros/src/lib.rs
@@ -12,300 +12,62 @@
  */
 
 use proc_macro::TokenStream;
-use quote::quote;
-use syn::{
-    Ident, ItemFn, Path, ReturnType, Token, Type,
-    parse::{Parse, ParseStream},
-    parse_macro_input,
-    punctuated::Punctuated,
-    visit::Visit,
-};
 
-/// Parses `name = <ident>` from the attribute arguments.
-struct Args {
-    name: Ident,
+mod shared;
+mod vendor_overridable;
+mod vendor_override;
+mod vendor_override_registry;
+
+fn render(result: syn::Result<proc_macro2::TokenStream>) -> TokenStream {
+    result.unwrap_or_else(syn::Error::into_compile_error).into()
 }
 
-impl Parse for Args {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let key: Ident = input.parse()?;
-        if key != "name" {
-            return Err(syn::Error::new(key.span(), "expected `name = <ident>`"));
-        }
-        input.parse::<Token![=]>()?;
-        let name: Ident = input.parse()?;
-        Ok(Args { name })
-    }
-}
-
-/// Visitor that checks whether a type mentions any of the given generic
-/// parameter identifiers (type or const parameters).
+/// Generates a `linkme` distributed-slice hook and public dispatcher from a
+/// fallback function.
 ///
-/// Any path whose *first* segment matches a generic parameter counts as a
-/// mention. This covers plain uses (`S`), associated types (`S::Assoc`), and
-/// const parameters (`N`), since inside the function signature a
-/// single-identifier first segment can only resolve to the generic parameter
-/// (generics shadow outer items of the same name).
-struct MentionsGenerics<'a> {
-    generics: &'a [Ident],
-    found: bool,
-}
-
-impl<'ast> Visit<'ast> for MentionsGenerics<'_> {
-    fn visit_path(&mut self, path: &'ast syn::Path) {
-        if path.leading_colon.is_none()
-            && let Some(seg) = path.segments.first()
-            && self.generics.contains(&seg.ident)
-        {
-            self.found = true;
-        }
-        syn::visit::visit_path(self, path);
-    }
-}
-
-fn mentions_generics(ty: &Type, generics: &[Ident]) -> bool {
-    let mut visitor = MentionsGenerics {
-        generics,
-        found: false,
-    };
-    visitor.visit_type(ty);
-    visitor.found
-}
-
-/// Extracts the plain identifier from a typed function parameter.
-///
-/// The generated dispatcher and shim code forward parameters by name, so
-/// patterns other than a plain identifier (`mut x`, `ref x`, tuple or struct
-/// destructuring, `x @ ..`) cannot be supported and are rejected with a
-/// compile error.
-fn param_ident(pat_type: &syn::PatType) -> syn::Result<Ident> {
-    match pat_type.pat.as_ref() {
-        syn::Pat::Ident(pi)
-            if pi.by_ref.is_none() && pi.mutability.is_none() && pi.subpat.is_none() =>
-        {
-            Ok(pi.ident.clone())
-        }
-        pat => Err(syn::Error::new_spanned(
-            pat,
-            "parameters must be plain identifiers (no `mut`, `ref`, or pattern destructuring)",
-        )),
-    }
-}
-
-/// Derives the hook static name from the public dispatcher name:
-/// e.g. `do_something` -> `DO_SOMETHING_HOOK`.
-///
-/// This naming scheme is part of the stable vendor-override contract; see
-/// [`macro@vendor_overridable`].
-fn hook_ident_for(name: &Ident) -> Ident {
-    Ident::new(
-        &format!("{}_HOOK", name.to_string().to_uppercase()),
-        name.span(),
-    )
-}
-
-/// Derives the paths of the crate-root vendor-override registry items
-/// (`__VENDOR_OVERRIDE_REGISTRY` and `__VendorOverrideRegistration`,
-/// generated by [`macro@declare_vendor_override_registry`]) from a hook
-/// path, by taking its first segment (`crate`, or the defining crate's
-/// name).
-///
-/// Rejects `self::...` / `super::...` paths and bare single-segment paths,
-/// since the registry lookup requires an explicit `crate` or crate-name
-/// prefix to resolve unambiguously from any module.
-fn registry_paths(
-    hook_path: &Path,
-) -> syn::Result<(proc_macro2::TokenStream, proc_macro2::TokenStream)> {
-    if hook_path.segments.len() < 2 {
-        return Err(syn::Error::new_spanned(
-            hook_path,
-            "vendor_override: hook path must include an explicit crate segment, e.g. \
-             `crate::name` or `crate_name::name`",
-        ));
-    }
-    let Some(first_seg) = hook_path.segments.first() else {
-        return Err(syn::Error::new_spanned(
-            hook_path,
-            "vendor_override: expected a non-empty path",
-        ));
-    };
-    if first_seg.ident == "self" || first_seg.ident == "super" {
-        return Err(syn::Error::new_spanned(
-            first_seg,
-            "vendor_override: hook path must be written as `crate::...` or `crate_name::...` (not \
-             `self::` or `super::`)",
-        ));
-    }
-    let leading_colon = &hook_path.leading_colon;
-    let crate_root = &first_seg.ident;
-    Ok((
-        quote! { #leading_colon #crate_root::__VENDOR_OVERRIDE_REGISTRY },
-        quote! { #leading_colon #crate_root::__VendorOverrideRegistration },
-    ))
-}
-
-/// Derives the registration static's name from the override function's own
-/// identifier, e.g. `do_something_vendor` -> `__DO_SOMETHING_VENDOR_OVERRIDE_REG`.
-fn registration_ident_for(name: &Ident) -> Ident {
-    Ident::new(
-        &format!("__{}_OVERRIDE_REG", name.to_string().to_uppercase()),
-        name.span(),
-    )
-}
-
-/// Hook parameter types and forwarding call expressions for a
-/// `vendor_overridable` fallback function, with any generic parameters
-/// erased to `&dyn Any` / `&mut dyn Any`.
-struct HookParams {
-    hook_param_types: Vec<proc_macro2::TokenStream>,
-    hook_call_args: Vec<proc_macro2::TokenStream>,
-    fallback_call_args: Vec<proc_macro2::TokenStream>,
-    any_erased: bool,
-}
-
-/// Builds the hook parameter types and forwarding call expressions for a
-/// `vendor_overridable` fallback function's parameters, erasing any
-/// parameter whose type mentions a generic parameter to `&dyn Any` /
-/// `&mut dyn Any`.
-///
-/// Returns `HookParams`.
-fn build_hook_params(
-    inputs: &Punctuated<syn::FnArg, Token![,]>,
-    generic_idents: &[Ident],
-) -> syn::Result<HookParams> {
-    let mut hook_param_types = Vec::new();
-    let mut hook_call_args = Vec::new();
-    let mut fallback_call_args = Vec::new();
-    let mut any_erased = false;
-
-    for arg in inputs {
-        let pat_type = match arg {
-            syn::FnArg::Typed(pat_type) => pat_type,
-            syn::FnArg::Receiver(_) => {
-                return Err(syn::Error::new_spanned(
-                    arg,
-                    "vendor_overridable does not support `self` parameters",
-                ));
-            }
-        };
-        let ident = param_ident(pat_type)?;
-        let ty = &pat_type.ty;
-        fallback_call_args.push(quote! { #ident });
-
-        if mentions_generics(ty, generic_idents) {
-            match ty.as_ref() {
-                Type::Reference(r) if r.mutability.is_none() => {
-                    any_erased = true;
-                    hook_param_types.push(quote! { &dyn ::core::any::Any });
-                    hook_call_args.push(quote! { #ident as &dyn ::core::any::Any });
-                }
-                Type::Reference(_) => {
-                    any_erased = true;
-                    hook_param_types.push(quote! { &mut dyn ::core::any::Any });
-                    hook_call_args.push(quote! { &mut *#ident as &mut dyn ::core::any::Any });
-                }
-                _ => {
-                    return Err(syn::Error::new_spanned(
-                        ty,
-                        "vendor_overridable: generic parameters must be references (`&Type<S>` or \
-                         `&mut Type<S>`) to be type-erasable",
-                    ));
-                }
-            }
-        } else {
-            hook_param_types.push(quote! { #ty });
-            hook_call_args.push(quote! { #ident });
-        }
-    }
-
-    Ok(HookParams {
-        hook_param_types,
-        hook_call_args,
-        fallback_call_args,
-        any_erased,
-    })
-}
-
-/// Forwards doc comments from the fallback function to the generated
-/// dispatcher, and builds the generated notes for the dispatcher and hook.
-///
-/// Returns `(doc_attrs, dispatcher_doc, hook_doc)`.
-fn fallback_docs<'a>(
-    fallback_fn: &'a ItemFn,
-    public_name: &Ident,
-    hook_ident: &Ident,
-) -> (Vec<&'a syn::Attribute>, String, String) {
-    let fallback_ident = &fallback_fn.sig.ident;
-    let doc_attrs: Vec<&syn::Attribute> = fallback_fn
-        .attrs
-        .iter()
-        .filter(|attr| attr.path().is_ident("doc"))
-        .collect();
-    let dispatcher_doc = format!(
-        "Vendor-overridable dispatcher generated by `vendor_overridable` from \
-         [`{fallback_ident}`]. Vendor overrides register through [`{hook_ident}`]."
-    );
-    let hook_doc = format!(
-        "Registration hook for vendor overrides of [`{public_name}`]. The name is derived from \
-         the dispatcher name (`{public_name}` -> `{hook_ident}`) and is part of the stable \
-         vendor-override contract."
-    );
-    (doc_attrs, dispatcher_doc, hook_doc)
-}
-
-/// Generates a `linkme` distributed-slice hook and a public dispatcher function
-/// from a fallback function.
+/// Generic parameters used by reference in input types are erased through
+/// `Any`. Return types may not reference generic parameters. Explicit lifetime
+/// parameters, `self` parameters, non-identifier parameter patterns, and async
+/// functions are unsupported.
 ///
 /// # Example
 /// ```rust,ignore
 /// #[override_macros::vendor_overridable(name = do_something)]
-/// fn do_something_fallback(param1: i32, param2: String) -> Result<(), String> {
-///     Ok(())
+/// fn do_something_fallback(value: u32) -> u32 {
+///     value
 /// }
 /// ```
 ///
-/// Expands to:
-/// - the original fallback function (unchanged)
-/// - `pub static DO_SOMETHING_HOOK: [fn(i32, String) -> Result<(), String>]` (linkme slice)
-/// - `pub fn do_something(param1: i32, param2: String) -> Result<(), String> { ... }`
-///
-/// Doc comments on the fallback function are forwarded to the generated
-/// dispatcher.
+/// This emits the unchanged fallback, a public `DO_SOMETHING_HOOK` distributed
+/// slice, and the public `do_something` dispatcher. Documentation from the
+/// fallback is forwarded to the dispatcher.
 ///
 /// # Hook naming contract
 ///
-/// The hook static's name is derived from the dispatcher name given via
-/// `name = ...` by upper-casing it and appending `_HOOK`
-/// (e.g. `do_something` -> `DO_SOMETHING_HOOK`). [`macro@vendor_override`]
+/// Hook names are derived by upper-casing the dispatcher name and appending
+/// `_HOOK` (`do_something` -> `DO_SOMETHING_HOOK`). [`macro@vendor_override`]
 /// relies on this derivation, and vendor crates link against the generated
-/// static by name: renaming the dispatcher is a breaking change for all
-/// vendor overrides.
+/// static by name. Renaming the dispatcher is therefore a breaking change for
+/// its vendor overrides.
 ///
 /// # Override registration
 ///
-/// At most one override may be registered per overridable function (generic
-/// or not). This is enforced two ways:
-/// - the crate that defines this function must call
-///   [`macro@declare_vendor_override_registry`] once at its crate root.
-///   Every [`macro@vendor_override`] attachment then self-registers into
-///   that crate's registry with no manual bookkeeping required, and the
-///   final binary calls the generated `<crate>::validate_vendor_overrides()`
-///   once at startup, which fails loudly if more than one override is
-///   linked in for the same hook (`linkme` slice order is otherwise
-///   unspecified, so a silently-picked override would be link-order
-///   dependent)
-/// - additionally, `debug_assert!` in the dispatcher itself, as a backstop
-///   for tests and binaries that forgot to call the validator
+/// At most one override may be registered for each dispatcher. The defining
+/// crate must invoke [`macro@declare_vendor_override_registry`] once at crate
+/// root. Every [`macro@vendor_override`] then self-registers in that crate's
+/// registry. Final binary must explicitly call generated
+/// `<crate>::validate_vendor_overrides()` during startup; it reports duplicate
+/// linked overrides, whose `linkme` ordering would otherwise be unspecified.
+/// Dispatcher also contains a `debug_assert!` as a backstop for tests and
+/// binaries that omit startup validation.
 ///
 /// # Generic fallback functions
 ///
-/// If the fallback function has generic type parameters, any parameter whose
-/// type mentions a generic (e.g. `&EcuManager<S>`) is *type-erased* to
-/// `&dyn core::any::Any` (or `&mut dyn core::any::Any` for `&mut` parameters)
-/// in the hook signature, while the public dispatcher keeps the full generics.
-/// The hook then returns `Option<Ret>` so an override can signal a failed
-/// downcast, in which case the dispatcher falls back to the default
-/// implementation:
+/// Parameters whose types mention generic parameters are type-erased to
+/// `&dyn core::any::Any` or `&mut dyn core::any::Any` in the hook, while the
+/// dispatcher preserves the original generic signature. Such hooks return
+/// `Option<Ret>`: a concrete override returns `None` when a downcast fails, and
+/// dispatcher then invokes the fallback.
 ///
 /// ```rust,ignore
 /// #[override_macros::vendor_overridable(name = do_something)]
@@ -314,322 +76,56 @@ fn fallback_docs<'a>(
 /// }
 /// ```
 ///
-/// Expands to:
-/// - `pub static DO_SOMETHING_HOOK: [fn(&dyn core::any::Any, &str) -> Option<u32>]`
-/// - `pub fn do_something<S: SomeTrait>(mgr: &Manager<S>, level: &str) -> u32 { ... }`
+/// This generates a hook equivalent to
+/// `fn(&dyn core::any::Any, &str) -> Option<u32>` and retains the generic
+/// dispatcher signature.
 ///
 /// Restrictions for generic fallbacks:
-/// - erased parameters must be references (`&Type<S>` or `&mut Type<S>`);
-///   by-value generic parameters are rejected at compile time
-/// - the return type must not mention any generic parameter
-/// - the erased type must be `'static` (required by `dyn Any`)
-/// - non-erased by-value parameters must be `Copy`, because the dispatcher
-///   may pass them to the override before falling back
-///
-/// The override registered via [`macro@vendor_override`] downcasts the
-/// erased reference back to the concrete type. If the concrete type does not
-/// match, the override returns `None` and the dispatcher falls back to the
-/// default implementation.
+/// - erased parameters must be references; generic values passed by value are
+///   rejected
+/// - return type must not mention generic parameters
+/// - erased concrete type must be `'static`, as required by `dyn Any`
+/// - non-erased values must remain usable if override declines and fallback is
+///   called; by-value parameters therefore generally need to be `Copy`
 ///
 /// # General restrictions
 ///
-/// - parameters must be plain identifiers (no `mut`, `ref`, or pattern
-///   destructuring)
-/// - explicit lifetime parameters are not supported (the hook is a `static`
-///   and cannot name them); elided lifetimes are fine
-/// - `self` parameters are not supported
+/// Async functions, explicit lifetime parameters, `self` parameters, and
+/// parameter patterns other than plain identifiers are unsupported. Elided
+/// lifetimes are supported.
 ///
 /// # Visibility requirement
 ///
-/// Vendor overriding is inherently a cross-crate mechanism: another crate
-/// must be able to name the generated hook static (via [`macro@vendor_override`]).
-/// The generated dispatcher function and hook static therefore carry
-/// `#[deny(unreachable_pub)]`. If the annotated function's module is not
-/// itself reachable from the crate root, add a `pub use` re-export (e.g. in
-/// `lib.rs`) to make the generated items reachable; otherwise the crate
-/// fails to compile with an "unreachable `pub` item" error.
+/// Vendor overrides are cross-crate. Generated dispatcher and hook carry
+/// `#[deny(unreachable_pub)]`; annotated function's module must be reachable
+/// from crate root, directly or through a `pub use` re-export.
 #[proc_macro_attribute]
 pub fn vendor_overridable(attr: TokenStream, item: TokenStream) -> TokenStream {
-    let args = parse_macro_input!(attr as Args);
-    let fallback_fn = parse_macro_input!(item as ItemFn);
-
-    let public_name = &args.name;
-    let hook_ident = hook_ident_for(public_name);
-
-    let fallback_ident = &fallback_fn.sig.ident;
-    let inputs = &fallback_fn.sig.inputs;
-    let output = &fallback_fn.sig.output;
-    let generics = &fallback_fn.sig.generics;
-    let where_clause = &generics.where_clause;
-
-    // The hook is a static; it cannot reference lifetime parameters of the
-    // fallback function.
-    if let Some(lifetime) = generics.lifetimes().next() {
-        return syn::Error::new_spanned(
-            lifetime,
-            "vendor_overridable: explicit lifetime parameters are not supported (the generated \
-             hook static cannot name them); use elided lifetimes instead",
-        )
-        .to_compile_error()
-        .into();
-    }
-
-    let generic_idents: Vec<Ident> = generics
-        .type_params()
-        .map(|tp| tp.ident.clone())
-        .chain(generics.const_params().map(|cp| cp.ident.clone()))
-        .collect();
-
-    let ret_ty = match output {
-        ReturnType::Default => quote! { () },
-        ReturnType::Type(_, ty) => {
-            if mentions_generics(ty, &generic_idents) {
-                return syn::Error::new_spanned(
-                    ty,
-                    "vendor_overridable: return type must not mention generic parameters",
-                )
-                .to_compile_error()
-                .into();
-            }
-            quote! { #ty }
-        }
-    };
-
-    // Build hook fn-ptr parameter types and dispatcher forwarding expressions.
-    // Parameters whose type mentions a generic are erased to `&dyn Any` /
-    // `&mut dyn Any`.
-    let HookParams {
-        hook_param_types,
-        hook_call_args,
-        fallback_call_args,
-        any_erased,
-    } = match build_hook_params(inputs, &generic_idents) {
-        Ok(v) => v,
-        Err(err) => return err.to_compile_error().into(),
-    };
-
-    let (doc_attrs, dispatcher_doc, hook_doc) =
-        fallback_docs(&fallback_fn, public_name, &hook_ident);
-
-    // Erased hooks return `Option<Ret>` so a shim can signal a failed downcast
-    // and let the dispatcher try the next override or fall back to the default
-    // implementation.
-    let expanded = if any_erased {
-        quote! {
-            #fallback_fn
-
-            #[::linkme::distributed_slice]
-            #[deny(unreachable_pub, reason = "vendor_overridable items must be reachable from \
-                the crate root so external crates can register overrides")]
-            #[doc = #hook_doc]
-            pub static #hook_ident: [fn(#(#hook_param_types),*) -> ::core::option::Option<#ret_ty>];
-
-            #(#doc_attrs)*
-            #[doc = ""]
-            #[doc = #dispatcher_doc]
-            #[deny(unreachable_pub, reason = "vendor_overridable items must be reachable from \
-                the crate root so external crates can register overrides")]
-            pub fn #public_name #generics (#inputs) -> #ret_ty #where_clause {
-                debug_assert!(
-                    #hook_ident.len() <= 1,
-                    "Multiple vendor overrides registered for `{}`; at most one override is \
-                     supported per overridable function",
-                    stringify!(#public_name),
-                );
-                if let Some(overridden_fn) = #hook_ident.first()
-                    && let Some(result) = overridden_fn(#(#hook_call_args),*)
-                {
-                    return result;
-                }
-                #fallback_ident(#(#fallback_call_args),*)
-            }
-        }
-    } else {
-        quote! {
-            #fallback_fn
-
-            #[::linkme::distributed_slice]
-            #[deny(unreachable_pub, reason = "vendor_overridable items must be reachable from \
-                the crate root so external crates can register overrides")]
-            #[doc = #hook_doc]
-            pub static #hook_ident: [fn(#(#hook_param_types),*) -> #ret_ty];
-
-            #(#doc_attrs)*
-            #[doc = ""]
-            #[doc = #dispatcher_doc]
-            #[deny(unreachable_pub, reason = "vendor_overridable items must be reachable from \
-                the crate root so external crates can register overrides")]
-            pub fn #public_name #generics (#inputs) -> #ret_ty #where_clause {
-                debug_assert!(
-                    #hook_ident.len() <= 1,
-                    "Multiple vendor overrides registered for `{}`; at most one override is \
-                     supported per overridable function",
-                    stringify!(#public_name),
-                );
-                if let Some(overridden_fn) = #hook_ident.first() {
-                    return overridden_fn(#(#hook_call_args),*);
-                }
-                #fallback_ident(#(#fallback_call_args),*)
-            }
-        }
-    };
-
-    expanded.into()
+    render(vendor_overridable::expand(attr.into(), item.into()))
 }
 
-/// Parses `path::to::fn_name` optionally followed by `, erase(param, ...)`.
-struct OverrideArgs {
-    hook_path: Path,
-    erased: Vec<Ident>,
-}
-
-impl Parse for OverrideArgs {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let hook_path: Path = input.parse()?;
-        let mut erased = Vec::new();
-        if input.peek(Token![,]) {
-            input.parse::<Token![,]>()?;
-            let key: Ident = input.parse()?;
-            if key != "erase" {
-                return Err(syn::Error::new(key.span(), "expected `erase(param, ...)`"));
-            }
-            let content;
-            syn::parenthesized!(content in input);
-            let params = Punctuated::<Ident, Token![,]>::parse_terminated(&content)?;
-            erased = params.into_iter().collect();
-        }
-        Ok(OverrideArgs { hook_path, erased })
-    }
-}
-
-/// Per-parameter pieces needed to build a `vendor_override` shim function:
-/// the shim's own parameter list, the `let ... else` downcast bindings for
-/// erased parameters, the arguments used to call the real override
-/// function, and the parameter types used in the compile-time signature
-/// assertion.
-struct ShimParts {
-    shim_params: Vec<proc_macro2::TokenStream>,
-    shim_bindings: Vec<proc_macro2::TokenStream>,
-    shim_call_args: Vec<proc_macro2::TokenStream>,
-    assert_param_types: Vec<proc_macro2::TokenStream>,
-}
-
-/// Builds the [`ShimParts`] for a `vendor_override` function, validating
-/// that every name listed in `erase(...)` matches a plain-identifier
-/// parameter and that erased parameters are references.
-fn build_shim_parts(func: &ItemFn, erased: &[Ident], fn_name: &str) -> syn::Result<ShimParts> {
-    let param_names: Vec<String> = func
-        .sig
-        .inputs
-        .iter()
-        .filter_map(|arg| match arg {
-            syn::FnArg::Typed(pat_type) => match pat_type.pat.as_ref() {
-                syn::Pat::Ident(pi) => Some(pi.ident.to_string()),
-                _ => None,
-            },
-            syn::FnArg::Receiver(_) => None,
-        })
-        .collect();
-    for name in erased {
-        if !param_names.iter().any(|p| name == p) {
-            return Err(syn::Error::new(
-                name.span(),
-                format!("vendor_override: no parameter named `{name}` in `{fn_name}`"),
-            ));
-        }
-    }
-
-    let mut parts = ShimParts {
-        shim_params: Vec::new(),
-        shim_bindings: Vec::new(),
-        shim_call_args: Vec::new(),
-        assert_param_types: Vec::new(),
-    };
-
-    for arg in &func.sig.inputs {
-        let pat_type = match arg {
-            syn::FnArg::Typed(pat_type) => pat_type,
-            syn::FnArg::Receiver(_) => {
-                return Err(syn::Error::new_spanned(
-                    arg,
-                    "vendor_override does not support `self` parameters",
-                ));
-            }
-        };
-        let ident = param_ident(pat_type)?;
-        let ty = &pat_type.ty;
-        parts.assert_param_types.push(quote! { #ty });
-
-        if erased.contains(&ident) {
-            let Type::Reference(ty_ref) = ty.as_ref() else {
-                return Err(syn::Error::new_spanned(
-                    ty,
-                    "vendor_override: erased parameters must be references (`&Type` or `&mut \
-                     Type`)",
-                ));
-            };
-            let concrete_ty = &ty_ref.elem;
-            let (any_ty, downcast) = if ty_ref.mutability.is_some() {
-                (
-                    quote! { &mut dyn ::core::any::Any },
-                    quote! { downcast_mut },
-                )
-            } else {
-                (quote! { &dyn ::core::any::Any }, quote! { downcast_ref })
-            };
-            parts.shim_params.push(quote! { #ident: #any_ty });
-            parts.shim_bindings.push(quote! {
-                let Some(#ident) = #ident.#downcast::<#concrete_ty>() else {
-                    ::tracing::debug!(
-                        "Vendor override `{}`: downcast of `{}` to `{}` failed, falling back to \
-                         the default implementation",
-                        #fn_name,
-                        stringify!(#ident),
-                        stringify!(#concrete_ty),
-                    );
-                    return ::core::option::Option::None;
-                };
-            });
-            parts.shim_call_args.push(quote! { #ident });
-        } else {
-            parts.shim_params.push(quote! { #ident: #ty });
-            parts.shim_call_args.push(quote! { #ident });
-        }
-    }
-
-    Ok(parts)
-}
-
-/// Generates the `#[linkme::distributed_slice(...)]` attribute for a vendor
-/// override function, deriving the hook static path from the overridable function path.
+/// Registers a function as vendor override for an overridable dispatcher.
 ///
-/// The hook static's name is derived from the last path segment by
-/// upper-casing it and appending `_HOOK` (the same derivation used by
-/// [`macro@vendor_overridable`], which documents the stable naming contract).
+/// Generic dispatcher parameters must be listed in `erase(...)` on the
+/// concrete override. Hook paths require an explicit `crate::` or crate-name
+/// prefix. Async override functions are unsupported.
 ///
 /// # Example
 /// ```rust,ignore
 /// #[override_macros::vendor_override(crate_a::do_something)]
-/// fn do_something_vendor(param1: i32, param2: String) -> Result<(), String> {
-///     Ok(())
+/// fn do_something_vendor(value: u32) -> u32 {
+///     value + 1
 /// }
 /// ```
 ///
-/// Expands to:
-/// ```rust,ignore
-/// #[linkme::distributed_slice(crate_a::DO_SOMETHING_HOOK)]
-/// fn do_something_vendor(param1: i32, param2: String) -> Result<(), String> {
-///     Ok(())
-/// }
-/// ```
+/// This derives the hook static from the final path segment using the stable
+/// naming contract documented by [`macro@vendor_overridable`], then registers
+/// the function in that `linkme` slice.
 ///
 /// # Overriding generic functions
 ///
-/// When the overridable function is generic, the hook stores type-erased
-/// `&dyn core::any::Any` / `&mut dyn core::any::Any` parameters and returns
-/// `Option<Ret>` (see [`macro@vendor_overridable`]). To override such a
-/// function, write the override with concrete types and list the erased
-/// parameters in `erase(...)`:
+/// Write generic overrides with concrete parameter types and list every
+/// type-erased parameter in `erase(...)`:
 ///
 /// ```rust,ignore
 /// #[override_macros::vendor_override(crate_a::do_something, erase(mgr))]
@@ -638,209 +134,51 @@ fn build_shim_parts(func: &ItemFn, erased: &[Ident], fn_name: &str) -> syn::Resu
 /// }
 /// ```
 ///
-/// This generates:
-/// - a hidden shim registered in the hook slice which downcasts each erased
-///   parameter to the concrete type declared by the override and returns
-///   `Some(result)`. If a downcast fails (the runtime generic type differs
-///   from the type declared by the override), the shim logs via
-///   `tracing::debug!` and returns `None`, causing the dispatcher to fall
-///   back to the default implementation. The overriding crate therefore
-///   needs a `tracing` dependency.
-/// - a compile-time assertion that the override signature is a valid
-///   instantiation of the overridable function (parameter types, return
-///   type, and trait bounds are checked against the public dispatcher).
+/// Macro emits original override plus a private shim. Shim downcasts each
+/// erased argument to declared concrete type and returns `Some(result)`. A
+/// failed downcast logs through `tracing::debug!` and returns `None`, causing
+/// dispatcher to use fallback. Overriding crate therefore needs `tracing`.
+/// Macro also emits a compile-time assertion that concrete parameter types,
+/// return type, and trait bounds form a valid dispatcher instantiation.
 ///
-/// This attribute also self-registers the override into the defining
-/// crate's override registry (see
-/// [`macro@declare_vendor_override_registry`]), so the crate-level
-/// `validate_vendor_overrides()` check can detect duplicate overrides
-/// without any manual bookkeeping at the attachment site. The registry is
-/// located via the hook path's crate segment, so hook paths must be written
-/// as `crate::...` or `crate_name::...` (not `self::...` / `super::...`,
-/// and not a bare single-segment path).
-///
-/// At most one override may be registered per overridable function, whether
-/// generic or not; see [`macro@vendor_overridable`] and
-/// [`macro@declare_vendor_override_registry`].
+/// Each attachment self-registers in defining crate's registry for duplicate
+/// detection. Hook path's first segment locates that registry, so paths must be
+/// `crate::...` or `crate_name::...`, never bare, `self::...`, or `super::...`.
+/// Final binary must call defining crate's `validate_vendor_overrides()` during
+/// startup. At most one override may target each dispatcher.
 #[proc_macro_attribute]
 pub fn vendor_override(attr: TokenStream, item: TokenStream) -> TokenStream {
-    let args = parse_macro_input!(attr as OverrideArgs);
-    // Keep the original function path for the compile-time signature assertion.
-    let public_fn_path = args.hook_path.clone();
-    let mut hook_path = args.hook_path;
-
-    // Replace the last path segment with the derived hook static name.
-    // e.g. `do_something` -> `DO_SOMETHING_HOOK`
-    let Some(last_seg) = hook_path.segments.last_mut() else {
-        return syn::Error::new_spanned(&hook_path, "vendor_override: expected a non-empty path")
-            .to_compile_error()
-            .into();
-    };
-    last_seg.ident = hook_ident_for(&last_seg.ident);
-
-    let (registry_path, reg_ty_path) = match registry_paths(&hook_path) {
-        Ok(v) => v,
-        Err(err) => return err.to_compile_error().into(),
-    };
-
-    let func = parse_macro_input!(item as ItemFn);
-    let fn_ident = &func.sig.ident;
-    let reg_ident = registration_ident_for(fn_ident);
-    let registration = quote! {
-        #[::linkme::distributed_slice(#registry_path)]
-        #[doc(hidden)]
-        static #reg_ident: #reg_ty_path = #reg_ty_path {
-            hook_name: stringify!(#public_fn_path),
-            count: || #hook_path.len(),
-        };
-    };
-
-    if args.erased.is_empty() {
-        let expanded = quote! {
-            #[::linkme::distributed_slice(#hook_path)]
-            #func
-
-            #registration
-        };
-        return expanded.into();
-    }
-
-    // Adapter generation: emit the override function unchanged plus a shim
-    // that downcasts erased parameters and is registered in the hook slice.
-    let fn_name = fn_ident.to_string();
-    let shim_ident = Ident::new(
-        &format!("__{fn_name}_vendor_override_shim"),
-        fn_ident.span(),
-    );
-    let output = &func.sig.output;
-    let ret_ty = match output {
-        ReturnType::Default => quote! { () },
-        ReturnType::Type(_, ty) => quote! { #ty },
-    };
-
-    let ShimParts {
-        shim_params,
-        shim_bindings,
-        shim_call_args,
-        assert_param_types,
-    } = match build_shim_parts(&func, &args.erased, &fn_name) {
-        Ok(parts) => parts,
-        Err(err) => return err.to_compile_error().into(),
-    };
-
-    let expanded = quote! {
-        #func
-
-        // Compile-time check: the override signature must be a valid
-        // instantiation of the overridable function.
-        const _: fn(#(#assert_param_types),*) -> #ret_ty = #public_fn_path;
-
-        #[::linkme::distributed_slice(#hook_path)]
-        fn #shim_ident(#(#shim_params),*) -> ::core::option::Option<#ret_ty> {
-            #(#shim_bindings)*
-            ::core::option::Option::Some(#fn_ident(#(#shim_call_args),*))
-        }
-
-        #registration
-    };
-
-    expanded.into()
+    render(vendor_override::expand(attr.into(), item.into()))
 }
 
-/// Declares the vendor-override registry for the crate this macro is
-/// invoked in. Must be called exactly once, at the crate root (typically
-/// `lib.rs`), by any crate that defines one or more
-/// [`macro@vendor_overridable`] functions.
+/// Declares vendor-override registration state for the invoking crate.
 ///
-/// Expands to:
-/// - a hidden `linkme` distributed slice, into which every
-///   [`macro@vendor_override`] attachment targeting a hook defined in this
-///   crate self-registers (no manual bookkeeping required at the
-///   attachment site)
-/// - `pub fn validate_vendor_overrides() -> Result<(), Vec<String>>`, which
-///   checks that every overridable function defined in this crate has at
-///   most one registered override linked in
-///
-/// `linkme` distributed slices are assembled by the linker, so the number
-/// of overrides registered for a given hook cannot be checked at compile
-/// time; `validate_vendor_overrides()` is a runtime check meant to be
-/// called once during application startup, before any overridable function
-/// defined in this crate is used.
+/// Invoke exactly once at crate root in each crate defining overridable
+/// functions. Generated `validate_vendor_overrides()` must be called explicitly
+/// by the final binary during startup.
 ///
 /// # Example
 /// ```rust,ignore
-/// // In the crate root, e.g. `lib.rs`:
 /// override_macros::declare_vendor_override_registry!();
 /// ```
+///
+/// This emits a hidden `linkme` distributed slice into which each
+/// [`macro@vendor_override`] targeting this crate self-registers, plus:
+///
 /// ```rust,ignore
-/// // At application startup, once per crate that defines overridable
-/// // functions:
+/// pub fn validate_vendor_overrides() -> Result<(), Vec<String>>
+/// ```
+///
+/// `linkme` slices are assembled by linker, so duplicate registrations cannot
+/// be checked at compile time. Final binary must call validator once during
+/// startup, before invoking overridable functions:
+///
+/// ```rust,ignore
 /// if let Err(errors) = cda_core::validate_vendor_overrides() {
 ///     panic!("Vendor override configuration error(s): {errors:?}");
 /// }
 /// ```
 #[proc_macro]
 pub fn declare_vendor_override_registry(input: TokenStream) -> TokenStream {
-    parse_macro_input!(input as syn::parse::Nothing);
-
-    let expanded = quote! {
-        /// Registration entry for one `override_macros::vendor_override`
-        /// attachment, collected via a `linkme` distributed slice into
-        /// `__VENDOR_OVERRIDE_REGISTRY`.
-        ///
-        /// Generated by `declare_vendor_override_registry!()`; internal to the
-        /// vendor-override mechanism, not part of its public contract.
-        #[doc(hidden)]
-        pub struct __VendorOverrideRegistration {
-            pub hook_name: &'static str,
-            pub count: fn() -> usize,
-        }
-
-        /// Distributed slice collecting one entry per `vendor_override`
-        /// attachment targeting a `vendor_overridable` function defined in
-        /// this crate. See [`validate_vendor_overrides`].
-        #[doc(hidden)]
-        #[::linkme::distributed_slice]
-        #[deny(unreachable_pub, reason = "must be reachable from the crate root so \
-            `vendor_override` sites in other crates can register into it")]
-        pub static __VENDOR_OVERRIDE_REGISTRY: [__VendorOverrideRegistration];
-
-        /// Checks that every vendor-overridable function defined in this
-        /// crate has at most one registered override linked in.
-        ///
-        /// `linkme` distributed slices are assembled by the linker, so this
-        /// cannot be checked at compile time. Call this once during
-        /// application startup, before any vendor-overridable function
-        /// defined in this crate is used.
-        ///
-        /// # Errors
-        /// Returns one message per overridable function that has more than
-        /// one override linked in.
-        pub fn validate_vendor_overrides()
-        -> ::core::result::Result<(), ::std::vec::Vec<::std::string::String>> {
-            let mut errors: ::std::vec::Vec<::std::string::String> = ::std::vec::Vec::new();
-            let mut seen: ::std::vec::Vec<&str> = ::std::vec::Vec::new();
-            for reg in __VENDOR_OVERRIDE_REGISTRY.iter() {
-                if seen.contains(&reg.hook_name) {
-                    continue;
-                }
-                seen.push(reg.hook_name);
-                let count = (reg.count)();
-                if count > 1 {
-                    errors.push(::std::format!(
-                        "`{}`: {} vendor overrides registered; at most one is supported",
-                        reg.hook_name,
-                        count,
-                    ));
-                }
-            }
-            if errors.is_empty() {
-                ::core::result::Result::Ok(())
-            } else {
-                ::core::result::Result::Err(errors)
-            }
-        }
-    };
-
-    expanded.into()
+    render(vendor_override_registry::expand(input.into()))
 }
