+++ override-macros/src/vendor_override_registry.rs
@@ -1,0 +1,77 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Copyright (c) Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use quote::quote;
+use syn::parse2;
+
+pub(crate) fn expand(input: proc_macro2::TokenStream) -> syn::Result<proc_macro2::TokenStream> {
+    parse2::<syn::parse::Nothing>(input)?;
+
+    Ok(quote! {
+        /// Registration entry for one `override_macros::vendor_override`
+        /// attachment, collected in `__VENDOR_OVERRIDE_REGISTRY`.
+        ///
+        /// Generated by `declare_vendor_override_registry!()` and internal to
+        /// the vendor-override mechanism, not part of its public contract.
+        #[doc(hidden)]
+        pub struct __VendorOverrideRegistration {
+            pub hook_name: &'static str,
+            pub count: fn() -> usize,
+        }
+
+        /// Distributed slice containing each `vendor_override` attachment that
+        /// targets a `vendor_overridable` function defined in this crate.
+        ///
+        /// See [`validate_vendor_overrides`].
+        #[doc(hidden)]
+        #[::linkme::distributed_slice]
+        #[deny(unreachable_pub, reason = "must be reachable from the crate root so \
+            `vendor_override` sites in other crates can register into it")]
+        pub static __VENDOR_OVERRIDE_REGISTRY: [__VendorOverrideRegistration];
+
+        /// Checks that every vendor-overridable function defined in this crate
+        /// has at most one registered override linked in.
+        ///
+        /// `linkme` distributed slices are assembled by the linker, so this
+        /// cannot be checked at compile time. Call this once during application
+        /// startup, before using any vendor-overridable function in this crate.
+        ///
+        /// # Errors
+        /// Returns one message per overridable function with multiple linked
+        /// overrides.
+        pub fn validate_vendor_overrides()
+        -> ::core::result::Result<(), ::std::vec::Vec<::std::string::String>> {
+            let mut errors: ::std::vec::Vec<::std::string::String> = ::std::vec::Vec::new();
+            let mut seen: ::std::vec::Vec<&str> = ::std::vec::Vec::new();
+            for reg in __VENDOR_OVERRIDE_REGISTRY.iter() {
+                if seen.contains(&reg.hook_name) {
+                    continue;
+                }
+                seen.push(reg.hook_name);
+                let count = (reg.count)();
+                if count > 1 {
+                    errors.push(::std::format!(
+                        "`{}`: {} vendor overrides registered; at most one is supported",
+                        reg.hook_name,
+                        count,
+                    ));
+                }
+            }
+            if errors.is_empty() {
+                ::core::result::Result::Ok(())
+            } else {
+                ::core::result::Result::Err(errors)
+            }
+        }
+    })
+}
